Issue #SC-1407 update: changes to report bucket
1. All reports are available in "reports" container.
