+++ platform-modules/batch-models/src/test/resources/application.conf
@@ -85,7 +85,6 @@ course.metrics.es.index.cbatchstats.prefix="cbatchstats-"
 course.metrics.es.index.cbatch="cbatch"
 course.metrics.cassandra.sunbirdKeyspace="sunbird"
 course.metrics.cassandra.sunbirdCoursesKeyspace="sunbird_courses"
-course.metrics.cloud.provider="azure"
 course.metrics.cloud.container="course-metrics"
 admin.reports.cloud.container="reports"
 # Folder names within container
@@ -93,11 +92,10 @@ course.metrics.cloud.objectKey="reports/"
 assessment.metrics.cloud.objectKey="assessment-reports/"
 admin.metrics.cloud.objectKey="public/admin-user-reports/"
 # End of report folder names
-course.metrics.temp.dir="/tmp/course-reports"
+course.metrics.temp.dir="/Users/sunil/Downloads/course-reports"
 course.metrics.cassandra.input.consistency="QUORUM"
 assessment.metrics.temp.dir="/Users/admin/assessment-report"
 assessment.metrics.cassandra.input.consistency="QUORUM"
-assessment.metrics.cloud.provider="azure"
 assessment.metrics.content.index="compositesearch"
 assessment.metrics.es.alias="cbatch-assessment"
 assessment.metrics.es.index.prefix="cbatch-assessment-"
@@ -106,7 +104,8 @@ course.es.index.enabled=true
 assessment.metrics.bestscore.report=true // BestScore or Latst Updated Score	
 assessment.metrics.supported.contenttype="SelfAssess"
 
-
+reports_azure_storage_key=""
+reports_azure_storage_secret=""
 # for only testing uploads to blob store
 azure_storage_key=""
 azure_storage_secret=""
